Moving over to docker-sync for Mac support.
Unison is written in OCaml. OCaml 4.02 changed the serialization format
over 4.01 so Unison binaries compiled with different versions of OCaml
are not compatible! The PHP-FPM container is based on Jessie where OCaml
4.01 is available. Using docker-sync allows a newer Linux container to
be used for Unison with OCaml 4.02. There are almost no working binaries
still available for Mac using OCaml 4.01. The Windows binary for OCaml
4.02 is now included. So finally managed to find a combination that
works across all platforms. Egad!